fixed type annotation in task 1

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -5,7 +5,7 @@ import asyncio
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: int) -> List[int]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     """ Executes wait_random n-times with max_delay as argument
         Returns a sorted list without using the sort method
     """
